fix(ci): Make keystore decoding more robust
The previous script failed if the base64 secret included the -----BEGIN/END----- headers generated by certutil.

This change uses `sed` to strip the first and last lines of the secret's content before piping to `base64 --decode`, making the process resilient to this format. Also adds quotes around the echo'd variable for safety.